ai_model/requirements.txt
tensorflow>=2.13
numpy>=1.23
pillow>=10.0
tqdm>=4.66
